@@ -1,79 +1,87 @@
 {
-  "name": "spatie/laravel-medialibrary",
-  "description": "Associate files with Eloquent models",
-  "keywords": [
-    "spatie",
-    "laravel-medialibrary",
-    "media",
-    "conversion",
-    "images",
-    "downloads",
-    "cms",
-    "laravel"
-  ],
-  "homepage": "https://github.com/spatie/laravel-medialibrary",
-  "license": "MIT",
-  "authors": [
-    {
-      "name": "Freek Van der Herten",
-      "email": "freek@spatie.be",
-      "homepage": "https://murze.be",
-      "role": "Developer"
-    }
-  ],
-  "require": {
-    "php": "^7.2",
-    "ext-fileinfo": "*",
-    "illuminate/bus": "~5.8.35|^6.0|^7.0",
-    "illuminate/console": "~5.8.35|^6.0|^7.0",
-    "illuminate/database": "~5.8.35|^6.0|^7.0",
-    "illuminate/pipeline": "~5.8.35|^6.0|^7.0",
-    "illuminate/support": "~5.8.35|^6.0|^7.0",
-    "league/flysystem": "^1.0.8",
-    "maennchen/zipstream-php": "^1.0",
-    "spatie/image": "^1.4.0",
-    "spatie/pdf-to-image": "^2.0",
-    "spatie/temporary-directory": "^1.1",
-    "symfony/console": "^4.4|^5.0",
-    "ext-json": "*"
-  },
-  "require-dev": {
-    "ext-pdo_sqlite": "*",
-    "doctrine/dbal": "^2.5.2",
-    "guzzlehttp/guzzle": "^6.3",
-    "league/flysystem-aws-s3-v3": "^1.0.13",
-    "mockery/mockery": "^1.0",
-    "orchestra/testbench": "^3.8|^4.0|^5.0",
-    "phpunit/phpunit" : "^8.0",
-    "spatie/phpunit-snapshot-assertions": "^2.0"
-  },
-  "conflict": {
-    "php-ffmpeg/php-ffmpeg": "<0.6.1"
-  },
-  "suggest": {
-    "league/flysystem-aws-s3-v3": "Required to use AWS S3 file storage",
-    "php-ffmpeg/php-ffmpeg": "Required for generating video thumbnails"
-  },
-  "autoload": {
-    "psr-4": {
-      "Spatie\\MediaLibrary\\": "src"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "Spatie\\MediaLibrary\\Tests\\": "tests"
-    }
-  },
-  "extra": {
-    "laravel": {
-      "providers": [
-        "Spatie\\MediaLibrary\\MediaLibraryServiceProvider"
-      ]
-    }
-  },
-  "config": {
-    "sort-packages": true
-  },
-  "minimum-stability": "dev",
-  "prefer-stable": true
+    "name": "spatie/laravel-medialibrary",
+    "description": "Associate files with Eloquent models",
+    "keywords": [
+        "spatie",
+        "laravel-medialibrary",
+        "media",
+        "conversion",
+        "images",
+        "downloads",
+        "cms",
+        "laravel"
+    ],
+    "homepage": "https://github.com/spatie/laravel-medialibrary",
+    "license": "MIT",
+    "authors": [
+        {
+            "name": "Freek Van der Herten",
+            "email": "freek@spatie.be",
+            "homepage": "https://spatie.be",
+            "role": "Developer"
+        }
+    ],
+    "require": {
+        "php": "^7.4",
+        "ext-fileinfo": "*",
+        "ext-json": "*",
+        "illuminate/bus": "^6.18|^7.0|^8.0",
+        "illuminate/console": "^6.18|^7.0|^8.0",
+        "illuminate/database": "^6.18|^7.0|^8.0",
+        "illuminate/pipeline": "^6.18|^7.0|^8.0",
+        "illuminate/support": "^6.18|^7.0|^8.0",
+        "league/flysystem": "^1.0.64",
+        "maennchen/zipstream-php": "^1.0|^2.0",
+        "spatie/image": "^1.4.0",
+        "spatie/temporary-directory": "^1.1",
+        "symfony/console": "^4.4|^5.0"
+    },
+    "require-dev": {
+        "ext-pdo_sqlite": "*",
+        "ext-zip": "*",
+        "aws/aws-sdk-php": "^3.133.11",
+        "doctrine/dbal": "^2.5.2",
+        "guzzlehttp/guzzle": "^6.3|^7.0",
+        "league/flysystem-aws-s3-v3": "^1.0.23",
+        "mockery/mockery": "^1.3",
+        "orchestra/testbench": "^4.0|^5.0|^6.0",
+        "php-ffmpeg/php-ffmpeg": "^0.16.0",
+        "phpunit/phpunit": "^9.1",
+        "spatie/pdf-to-image": "^2.0",
+        "spatie/phpunit-snapshot-assertions": "^4.0"
+    },
+    "conflict": {
+        "php-ffmpeg/php-ffmpeg": "<0.6.1"
+    },
+    "suggest": {
+        "league/flysystem-aws-s3-v3": "Required to use AWS S3 file storage",
+        "php-ffmpeg/php-ffmpeg": "Required for generating video thumbnails",
+        "spatie/pdf-to-image": "Required for generating thumbsnails of PDFs and SVGs"
+    },
+    "autoload": {
+        "psr-4": {
+            "Spatie\\MediaLibrary\\": "src"
+        }
+    },
+    "scripts": {
+        "format": "vendor/bin/php-cs-fixer fix --allow-risky=yes",
+        "test": "vendor/bin/phpunit"
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Spatie\\MediaLibrary\\Tests\\": "tests"
+        }
+    },
+    "extra": {
+        "laravel": {
+            "providers": [
+                "Spatie\\MediaLibrary\\MediaLibraryServiceProvider"
+            ]
+        }
+    },
+    "config": {
+        "sort-packages": true
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
