@@ -2,42 +2,34 @@
 
 namespace Spatie\MediaLibrary\ResponsiveImages;
 
-use Spatie\MediaLibrary\Models\Media;
+use Illuminate\Support\Collection;
+use Spatie\MediaLibrary\MediaCollections\Models\Media;
 
 class RegisteredResponsiveImages
 {
-    /** Spatie\Medialibrary\Media */
-    protected $media;
+    protected Media $media;
 
-    /** Illuminate\Support\Collection */
-    public $files;
+    public Collection $files;
 
-    /** string */
-    public $generatedFor;
+    public string $generatedFor;
 
     public function __construct(Media $media, string $conversionName = '')
     {
         $this->media = $media;
 
         $this->generatedFor = $conversionName === ''
-            ? 'medialibrary_original'
+            ? 'media_library_original'
             : $conversionName;
 
         $this->files = collect($media->responsive_images[$this->generatedFor]['urls'] ?? [])
-            ->map(function (string $fileName) use ($media) {
-                return new ResponsiveImage($fileName, $media);
-            })
-            ->filter(function (ResponsiveImage $responsiveImage) {
-                return $responsiveImage->generatedFor() === $this->generatedFor;
-            });
+            ->map(fn (string $fileName) => new ResponsiveImage($fileName, $media))
+            ->filter(fn (ResponsiveImage $responsiveImage) => $responsiveImage->generatedFor() === $this->generatedFor);
     }
 
     public function getUrls(): array
     {
         return $this->files
-            ->map(function (ResponsiveImage $responsiveImage) {
-                return $responsiveImage->url();
-            })
+            ->map(fn (ResponsiveImage $responsiveImage) => $responsiveImage->url())
             ->values()
             ->toArray();
     }
@@ -45,12 +37,10 @@ class RegisteredResponsiveImages
     public function getSrcset(): string
     {
         $filesSrcset = $this->files
-            ->map(function (ResponsiveImage $responsiveImage) {
-                return "{$responsiveImage->url()} {$responsiveImage->width()}w";
-            })
+            ->map(fn (ResponsiveImage $responsiveImage) => "{$responsiveImage->url()} {$responsiveImage->width()}w")
             ->implode(', ');
 
-        $shouldAddPlaceholderSvg = config('medialibrary.responsive_images.use_tiny_placeholders')
+        $shouldAddPlaceholderSvg = config('media-library.responsive_images.use_tiny_placeholders')
             && $this->getPlaceholderSvg();
 
         if ($shouldAddPlaceholderSvg) {
