@@ -3,14 +3,15 @@
 namespace Spatie\MediaLibrary\Tests\Feature\FileAdder;
 
 use Illuminate\Support\Facades\Storage;
-use Spatie\MediaLibrary\Exceptions\FileCannotBeAdded\DiskDoesNotExist;
-use Spatie\MediaLibrary\Exceptions\FileCannotBeAdded\FileDoesNotExist;
-use Spatie\MediaLibrary\Exceptions\FileCannotBeAdded\FileIsTooBig;
-use Spatie\MediaLibrary\Exceptions\FileCannotBeAdded\InvalidBase64Data;
-use Spatie\MediaLibrary\Exceptions\FileCannotBeAdded\MimeTypeNotAllowed;
-use Spatie\MediaLibrary\Exceptions\FileCannotBeAdded\RequestDoesNotHaveFile;
-use Spatie\MediaLibrary\Exceptions\FileCannotBeAdded\UnknownType;
-use Spatie\MediaLibrary\Exceptions\FileCannotBeAdded\UnreachableUrl;
+use Spatie\MediaLibrary\MediaCollections\Exceptions\DiskDoesNotExist;
+use Spatie\MediaLibrary\MediaCollections\Exceptions\FileDoesNotExist;
+use Spatie\MediaLibrary\MediaCollections\Exceptions\FileIsTooBig;
+use Spatie\MediaLibrary\MediaCollections\Exceptions\InvalidBase64Data;
+use Spatie\MediaLibrary\MediaCollections\Exceptions\InvalidUrl;
+use Spatie\MediaLibrary\MediaCollections\Exceptions\MimeTypeNotAllowed;
+use Spatie\MediaLibrary\MediaCollections\Exceptions\RequestDoesNotHaveFile;
+use Spatie\MediaLibrary\MediaCollections\Exceptions\UnknownType;
+use Spatie\MediaLibrary\MediaCollections\Exceptions\UnreachableUrl;
 use Spatie\MediaLibrary\Tests\TestCase;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 
@@ -22,6 +23,16 @@ class IntegrationTest extends TestCase
         $media = $this->testModel
             ->addMedia($this->getTestJpg())
             ->toMediaCollection();
+
+        $this->assertEquals('default', $media->collection_name);
+    }
+
+    /** @test */
+    public function toMediaCollection_has_an_alias_called_toMediaLibrary()
+    {
+        $media = $this->testModel
+            ->addMedia($this->getTestJpg())
+            ->toMediaLibrary();
 
         $this->assertEquals('default', $media->collection_name);
     }
@@ -59,7 +70,7 @@ class IntegrationTest extends TestCase
     }
 
     /** @test */
-    public function it_can_move_the_original_file_to_the_medialibrary()
+    public function it_can_move_the_original_file_to_the_media_library()
     {
         $testFile = $this->getTestJpg();
 
@@ -67,12 +78,12 @@ class IntegrationTest extends TestCase
             ->addMedia($testFile)
             ->toMediaCollection();
 
-        $this->assertFileNotExists($testFile);
+        $this->assertFileDoesNotExist($testFile);
         $this->assertFileExists($this->getMediaDirectory($media->id.'/'.$media->file_name));
     }
 
     /** @test */
-    public function it_can_copy_the_original_file_to_the_medialibrary()
+    public function it_can_copy_the_original_file_to_the_media_library()
     {
         $testFile = $this->getTestJpg();
 
@@ -127,7 +138,7 @@ class IntegrationTest extends TestCase
     }
 
     /** @test */
-    public function it_can_add_an_upload_to_the_medialibrary()
+    public function it_can_add_an_upload_to_the_media_library()
     {
         $uploadedFile = new UploadedFile(
             $this->getTestFilesDirectory('test.jpg'),
@@ -145,7 +156,7 @@ class IntegrationTest extends TestCase
     }
 
     /** @test */
-    public function it_can_add_an_upload_to_the_medialibrary_from_the_current_request()
+    public function it_can_add_an_upload_to_the_media_library_from_the_current_request()
     {
         $this->app['router']->get('/upload', function () {
             $media = $this->testModel
@@ -169,7 +180,7 @@ class IntegrationTest extends TestCase
     }
 
     /** @test */
-    public function it_can_add_multiple_uploads_to_the_medialibrary_from_the_current_request()
+    public function it_can_add_multiple_uploads_to_the_media_library_from_the_current_request()
     {
         $this->app['router']->get('/upload', function () {
             $fileAdders = collect(
@@ -208,7 +219,7 @@ class IntegrationTest extends TestCase
     }
 
     /** @test */
-    public function it_can_add_handle_file_keys_that_contain_an_array_to_the_medialibrary_from_the_current_request()
+    public function it_can_add_handle_file_keys_that_contain_an_array_to_the_media_library_from_the_current_request()
     {
         $this->app['router']->get('/upload', function () {
             $fileAdders = collect(
@@ -284,9 +295,9 @@ class IntegrationTest extends TestCase
     }
 
     /** @test */
-    public function it_can_add_a_remote_file_to_the_medialibrary()
+    public function it_can_add_a_remote_file_to_the_media_library()
     {
-        $url = 'https://docs.spatie.be/laravel-medialibrary/v7/images/header.jpg';
+        $url = 'https://docs.spatie.be/laravel-medialibrary/v8/images/header.jpg';
 
         $media = $this->testModel
             ->addMediaFromUrl($url)
@@ -297,7 +308,17 @@ class IntegrationTest extends TestCase
     }
 
     /** @test */
-    public function it_can_add_a_file_from_a_separate_disk_to_the_medialibrary()
+    public function it_will_not_add_local_files_when_an_url_is_expected()
+    {
+        $this->expectException(InvalidUrl::class);
+
+        $this->testModel
+            ->addMediaFromUrl(__FILE__)
+            ->toMediaCollection();
+    }
+
+    /** @test */
+    public function it_can_add_a_file_from_a_separate_disk_to_the_media_library()
     {
         Storage::disk('secondMediaDisk')->put('tmp/test.jpg', file_get_contents($this->getTestJpg()));
 
@@ -309,43 +330,41 @@ class IntegrationTest extends TestCase
     }
 
     /** @test */
-    public function it_can_add_a_remote_file_with_the_name_of_the_last_directory_to_the_medialibrary()
+    public function it_can_natively_copy_a_remote_file_from_the_same_disk_to_the_media_library()
     {
-        $this->markTestSkipped('Find a different way to implement the url.');
-
-        $url = 'https://docs.spatie.be/laravel-medialibrary/test-image';
+        Storage::disk('public')->put('tmp/test.jpg', file_get_contents($this->getTestJpg()));
+        $this->assertFileExists($this->getMediaDirectory('tmp/test.jpg'));
 
         $media = $this->testModel
-            ->addMediaFromUrl($url)
+            ->addMediaFromDisk('tmp/test.jpg', 'public')
             ->toMediaCollection();
 
-        $this->assertFileExists($this->getMediaDirectory("{$media->id}/test-image.jpeg"));
+        $this->assertFileExists($this->getMediaDirectory("{$media->id}/test.jpg"));
+        $this->assertFileDoesNotExist($this->getMediaDirectory('tmp/test.jpg'));
     }
 
     /** @test */
-    public function it_can_add_a_remote_file_with_no_name_or_directory_to_the_medialibrary()
+    public function it_can_add_a_remote_file_with_a_space_in_the_name_to_the_media_library()
     {
-        $this->markTestSkipped('Find a different way to implement the url.');
-
-        $url = 'https://docs.spatie.be/?test-image=true';
-
-        $media = $this->testModel
-            ->addMediaFromUrl($url)
-            ->toMediaCollection();
-
-        $this->assertFileExists($this->getMediaDirectory("{$media->id}/file.jpeg"));
-    }
-
-    /** @test */
-    public function it_can_add_a_remote_file_with_a_space_in_the_name_to_the_medialibrary()
-    {
-        $url = 'http://spatie.github.io/laravel-medialibrary/tests/Support/testfiles/test%20with%20space.jpg';
+        $url = 'http://spatie.github.io/laravel-medialibrary/tests/TestSupport/testfiles/test%20with%20space.jpg';
 
         $media = $this->testModel
             ->addMediaFromUrl($url)
             ->toMediaCollection();
 
         $this->assertFileExists($this->getMediaDirectory("{$media->id}/test-with-space.jpg"));
+    }
+
+    /** @test */
+    public function it_can_add_a_remote_file_with_an_accent_in_the_name_to_the_media_library()
+    {
+        $url = 'https://orbit.brightbox.com/v1/acc-jqzwj/Marquis-Leisure/reviews/images/000/000/898/original/Antar%C3%A8sThumb.jpg';
+
+        $media = $this->testModel
+            ->addMediaFromUrl($url)
+            ->toMediaCollection();
+
+        $this->assertFileExists($this->getMediaDirectory("{$media->id}/AntarèsThumb.jpg"));
     }
 
     /** @test */
@@ -363,7 +382,7 @@ class IntegrationTest extends TestCase
     /** @test */
     public function it_wil_throw_an_exception_when_a_remote_file_has_an_invalid_mime_type()
     {
-        $url = 'https://docs.spatie.be/laravel-medialibrary/v7/images/header.jpg';
+        $url = 'https://docs.spatie.be/laravel-medialibrary/v8/images/header.jpg';
 
         $this->expectException(MimeTypeNotAllowed::class);
 
@@ -413,9 +432,7 @@ class IntegrationTest extends TestCase
     {
         $media = $this->testModel
             ->addMedia($this->getTestJpg())
-            ->sanitizingFileName(function ($fileName) {
-                return 'new_file_name.jpg';
-            })
+            ->sanitizingFileName(fn ($fileName) => 'new_file_name.jpg')
             ->toMediaCollection();
 
         $this->assertEquals('test', $media->name);
@@ -493,7 +510,7 @@ class IntegrationTest extends TestCase
     /** @test */
     public function it_will_throw_an_exception_when_adding_a_file_that_is_too_big()
     {
-        $this->app['config']->set('medialibrary.max_file_size', 1);
+        $this->app['config']->set('media-library.max_file_size', 1);
 
         $this->expectException(FileIsTooBig::class);
 
@@ -513,7 +530,7 @@ class IntegrationTest extends TestCase
     }
 
     /** @test */
-    public function it_can_add_a_base64_encoded_file_to_the_medialibrary()
+    public function it_can_add_a_base64_encoded_file_to_the_media_library()
     {
         $testFile = $this->getTestJpg();
         $testBase64Data = base64_encode(file_get_contents($testFile));
@@ -523,6 +540,18 @@ class IntegrationTest extends TestCase
             ->toMediaCollection();
 
         $this->assertFileExists($this->getMediaDirectory($media->id.'/'.$media->file_name));
+    }
+
+    /** @test */
+    public function a_string_can_be_accepted_to_be_added_to_the_media_library()
+    {
+        $string = 'test123';
+
+        $media = $this->testModel
+            ->addMediaFromString($string)
+            ->toMediaCollection();
+
+        $this->assertEquals($string, file_get_contents($media->getPath()));
     }
 
     /** @test */
@@ -581,7 +610,7 @@ class IntegrationTest extends TestCase
     }
 
     /** @test */
-    public function it_can_add_an_upload_to_the_medialibrary_using_dot_notation()
+    public function it_can_add_an_upload_to_the_media_library_using_dot_notation()
     {
         $this->app['router']->get('/upload', function () {
             $media = $this->testModel
