@@ -1,6 +1,6 @@
 <?php
 
-namespace Spatie\MediaLibrary\Tests\Feature\Models\Media;
+namespace Spatie\MediaLibrary\Tests\Feature\Media;
 
 use Spatie\MediaLibrary\Tests\TestCase;
 
@@ -9,11 +9,9 @@ class ResponsableTest extends TestCase
     /** @test */
     public function it_can_return_an_image_as_a_response()
     {
-        $this->app['router']->get('/upload', function () {
-            return $this->testModel
-                ->addMedia($this->getTestJpg())
-                ->toMediaCollection();
-        });
+        $this->app['router']->get('/upload', fn () => $this->testModel
+            ->addMedia($this->getTestJpg())
+            ->toMediaCollection());
 
         $result = $this->call('get', 'upload');
 
@@ -25,11 +23,9 @@ class ResponsableTest extends TestCase
     /** @test */
     public function it_can_return_a_text_as_a_response()
     {
-        $this->app['router']->get('/upload', function () {
-            return $this->testModel
-                ->addMedia($this->getTestFilesDirectory('test.txt'))
-                ->toMediaCollection();
-        });
+        $this->app['router']->get('/upload', fn () => $this->testModel
+            ->addMedia($this->getTestFilesDirectory('test.txt'))
+            ->toMediaCollection());
 
         $result = $this->call('get', 'upload');
 
