@@ -3,10 +3,12 @@
 namespace Spatie\MediaLibrary;
 
 use Illuminate\Support\ServiceProvider;
-use Spatie\MediaLibrary\Commands\CleanCommand;
-use Spatie\MediaLibrary\Commands\ClearCommand;
-use Spatie\MediaLibrary\Commands\RegenerateCommand;
-use Spatie\MediaLibrary\Filesystem\Filesystem;
+use Spatie\MediaLibrary\Conversions\Commands\RegenerateCommand;
+use Spatie\MediaLibrary\MediaCollections\Commands\CleanCommand;
+use Spatie\MediaLibrary\MediaCollections\Commands\ClearCommand;
+use Spatie\MediaLibrary\MediaCollections\Filesystem;
+use Spatie\MediaLibrary\MediaCollections\MediaRepository;
+use Spatie\MediaLibrary\MediaCollections\Models\Observers\MediaObserver;
 use Spatie\MediaLibrary\ResponsiveImages\TinyPlaceholderGenerator\TinyPlaceholderGenerator;
 use Spatie\MediaLibrary\ResponsiveImages\WidthCalculator\WidthCalculator;
 
@@ -14,13 +16,17 @@ class MediaLibraryServiceProvider extends ServiceProvider
 {
     public function boot()
     {
-        $this->publishes([
-            __DIR__.'/../config/medialibrary.php' => config_path('medialibrary.php'),
-        ], 'config');
+        $this->registerPublishables();
 
-        if (! class_exists('CreateMediaTable')) {
+        if (! class_exists('CreateModelHasMediaTable')) {
             $this->publishes([
-                __DIR__.'/../database/migrations/create_media_table.php.stub' => database_path('migrations/'.date('Y_m_d_His', time()).'_create_media_table.php'),
+                __DIR__.'/../database/migrations/create_model_has_media_table.php.stub' => database_path('migrations/'.date('Y_m_d_His', time()).'_create_model_has_media_table.php'),
+            ], 'migrations');
+        }
+
+        if (! class_exists('CreateModelHasMediaTable')) {
+            $this->publishes([
+                __DIR__.'/../database/migrations/create_model_has_media_table.php.stub' => database_path('migrations/'.date('Y_m_d_His', time()).'_create_model_has_media_table.php'),
             ], 'migrations');
         }
 
@@ -32,41 +38,53 @@ class MediaLibraryServiceProvider extends ServiceProvider
 
         $mediaClass::observe(new MediaObserver());
 
-        $this->loadViewsFrom(__DIR__.'/../resources/views', 'medialibrary');
+        $this->loadViewsFrom(__DIR__.'/../resources/views', 'media-library');
     }
 
     public function register()
     {
-        $this->mergeConfigFrom(__DIR__.'/../config/medialibrary.php', 'medialibrary');
+        $this->mergeConfigFrom(__DIR__.'/../config/media-library.php', 'media-library');
 
         $this->app->singleton(MediaRepository::class, function () {
-            $mediaClass = config('medialibrary.media_model');
+            $mediaClass = config('media-library.media_model');
 
             return new MediaRepository(new $mediaClass);
         });
 
-        $this->app->bind('command.medialibrary:regenerate', RegenerateCommand::class);
-        $this->app->bind('command.medialibrary:clear', ClearCommand::class);
-        $this->app->bind('command.medialibrary:clean', CleanCommand::class);
-
-        $this->app->bind(Filesystem::class, Filesystem::class);
-
-        $this->app->bind(WidthCalculator::class, config('medialibrary.responsive_images.width_calculator'));
-        $this->app->bind(TinyPlaceholderGenerator::class, config('medialibrary.responsive_images.tiny_placeholder_generator'));
-
-        $this->commands([
-            'command.medialibrary:regenerate',
-            'command.medialibrary:clear',
-            'command.medialibrary:clean',
-        ]);
-
-        $this->registerDeprecatedConfig();
+        $this->registerCommands();
     }
 
-    protected function registerDeprecatedConfig()
+    protected function registerPublishables(): void
     {
-        if (! config('medialibrary.disk_name')) {
-            config(['medialibrary.disk_name' => config('medialibrary.default_filesystem')]);
+        $this->publishes([
+            __DIR__.'/../config/media-library.php' => config_path('media-library.php'),
+        ], 'config');
+
+        if (! class_exists('CreateMediaTable')) {
+            $this->publishes([
+                __DIR__.'/../database/migrations/create_media_table.php.stub' => database_path('migrations/'.date('Y_m_d_His', time()).'_create_media_table.php'),
+            ], 'migrations');
         }
+
+        $this->publishes([
+            __DIR__.'/../resources/views' => resource_path('views/vendor/media-library'),
+        ], 'views');
+    }
+
+    protected function registerCommands(): void
+    {
+        $this->app->bind(Filesystem::class, Filesystem::class);
+        $this->app->bind(WidthCalculator::class, config('media-library.responsive_images.width_calculator'));
+        $this->app->bind(TinyPlaceholderGenerator::class, config('media-library.responsive_images.tiny_placeholder_generator'));
+
+        $this->app->bind('command.media-library:regenerate', RegenerateCommand::class);
+        $this->app->bind('command.media-library:clear', ClearCommand::class);
+        $this->app->bind('command.media-library:clean', CleanCommand::class);
+
+        $this->commands([
+            'command.media-library:regenerate',
+            'command.media-library:clear',
+            'command.media-library:clean',
+        ]);
     }
 }
