@@ -3,16 +3,16 @@
 namespace Spatie\MediaLibrary\ResponsiveImages\Exceptions;
 
 use Exception;
-use Spatie\MediaLibrary\Helpers\File;
+use Spatie\MediaLibrary\Support\File;
 
 class InvalidTinyJpg extends Exception
 {
-    public static function doesNotExist(string $tinyImageDestinationPath)
+    public static function doesNotExist(string $tinyImageDestinationPath): self
     {
         return new static("The expected tiny jpg at `{$tinyImageDestinationPath}` does not exist");
     }
 
-    public static function hasWrongMimeType(string $tinyImageDestinationPath)
+    public static function hasWrongMimeType(string $tinyImageDestinationPath): self
     {
         $foundMimeType = File::getMimeType($tinyImageDestinationPath);
 
