@@ -3,7 +3,7 @@
 namespace Spatie\MediaLibrary\ResponsiveImages\WidthCalculator;
 
 use Illuminate\Support\Collection;
-use Spatie\MediaLibrary\Helpers\ImageFactory;
+use Spatie\MediaLibrary\Support\ImageFactory;
 
 class FileSizeOptimizedWidthCalculator implements WidthCalculator
 {
