@@ -1,9 +1,9 @@
 <?php
 
-namespace Spatie\MediaLibrary\Tests\Feature\Models\Media;
+namespace Spatie\MediaLibrary\Tests\Feature\Media;
 
-use Spatie\MediaLibrary\Tests\Support\TestModels\TestModel;
 use Spatie\MediaLibrary\Tests\TestCase;
+use Spatie\MediaLibrary\Tests\TestSupport\TestModels\TestModel;
 
 class CopyTest extends TestCase
 {
@@ -13,7 +13,7 @@ class CopyTest extends TestCase
         /** @var TestModel $model */
         $model = TestModel::create(['name' => 'test']);
 
-        /** @var \Spatie\MediaLibrary\Models\Media $media */
+        /** @var \Spatie\MediaLibrary\MediaCollections\Models\Media $media */
         $media = $model
             ->addMedia($this->getTestJpg())
             ->usingName('custom-name')
@@ -41,16 +41,16 @@ class CopyTest extends TestCase
     /** @test */
     public function it_can_copy_file_without_extension()
     {
-        if (! file_exists(storage_path('medialibrary/temp'))) {
-            mkdir(storage_path('medialibrary/temp'), 0777, true);
+        if (! file_exists(storage_path('media-library/temp'))) {
+            mkdir(storage_path('media-library/temp'), 0777, true);
         }
 
-        config(['medialibrary.temporary_directory_path' => realpath(storage_path('medialibrary/temp'))]);
+        config(['media-library.temporary_directory_path' => realpath(storage_path('media-library/temp'))]);
 
         /** @var TestModel $model */
         $model = TestModel::create(['name' => 'test']);
 
-        /** @var \Spatie\MediaLibrary\Models\Media $media */
+        /** @var \Spatie\MediaLibrary\MediaCollections\Models\Media $media */
         $media = $model
             ->addMedia($this->getTestImageWithoutExtension())
             ->usingName('custom-name')
@@ -83,7 +83,7 @@ class CopyTest extends TestCase
         /** @var TestModel $model */
         $model = TestModel::create(['name' => 'test']);
 
-        /** @var \Spatie\MediaLibrary\Models\Media $media */
+        /** @var \Spatie\MediaLibrary\MediaCollections\Models\Media $media */
         $media = $model
             ->addMedia($this->getTestJpg())
             ->usingName('custom-name')
@@ -105,6 +105,43 @@ class CopyTest extends TestCase
         $this->assertFileExists($this->getTempDirectory('media2').'/'.$movedMedia->id.'/test.jpg');
         $this->assertEquals($movedMedia->collection_name, 'images');
         $this->assertEquals($movedMedia->disk, $diskName);
+        $this->assertEquals($movedMedia->model->id, $anotherModel->id);
+        $this->assertEquals($movedMedia->name, 'custom-name');
+        $this->assertEquals($movedMedia->getCustomProperty('custom-property-name'), 'custom-property-value');
+    }
+
+    /** @test */
+    public function it_can_copy_file_with_accent()
+    {
+        if (! file_exists(storage_path('media-library/temp'))) {
+            mkdir(storage_path('media-library/temp'), 0777, true);
+        }
+
+        config(['media-library.temporary_directory_path' => realpath(storage_path('media-library/temp'))]);
+
+        /** @var TestModel $model */
+        $model = TestModel::create(['name' => 'test']);
+
+        /** @var \Spatie\MediaLibrary\MediaCollections\Models\Media $media */
+        $media = $model
+            ->addMedia($this->getAntaresThumbJpgWithAccent())
+            ->usingName('custom-name')
+            ->withCustomProperties(['custom-property-name' => 'custom-property-value'])
+            ->toMediaCollection();
+
+        $this->assertFileExists($this->getMediaDirectory($media->id.'/antarèsthumb.jpg'));
+
+        $anotherModel = TestModel::create(['name' => 'another-test']);
+
+        $movedMedia = $media->copy($anotherModel, 'images');
+
+        $movedMedia->refresh();
+
+        $this->assertCount(1, $model->getMedia('default'));
+        $this->assertFileExists($this->getMediaDirectory($media->id.'/antarèsthumb.jpg'));
+
+        $this->assertCount(1, $anotherModel->getMedia('images'));
+        $this->assertFileExists($this->getMediaDirectory($movedMedia->id.'/antarèsthumb.jpg'));
         $this->assertEquals($movedMedia->model->id, $anotherModel->id);
         $this->assertEquals($movedMedia->name, 'custom-name');
         $this->assertEquals($movedMedia->getCustomProperty('custom-property-name'), 'custom-property-value');
