@@ -1,13 +1,13 @@
 <?php
 
-namespace Spatie\MediaLibrary\Tests\Feature\Models\Media;
+namespace Spatie\MediaLibrary\Tests\Feature\Media;
 
 use File;
 use Illuminate\Database\Eloquent\SoftDeletes;
-use Spatie\MediaLibrary\Models\Media;
-use Spatie\MediaLibrary\Tests\Support\TestModels\TestModel;
-use Spatie\MediaLibrary\Tests\Support\TestPathGenerator;
+use Spatie\MediaLibrary\MediaCollections\Models\Media;
 use Spatie\MediaLibrary\Tests\TestCase;
+use Spatie\MediaLibrary\Tests\TestSupport\TestModels\TestModel;
+use Spatie\MediaLibrary\Tests\TestSupport\TestPathGenerator;
 
 class DeleteTest extends TestCase
 {
@@ -38,7 +38,7 @@ class DeleteTest extends TestCase
     /** @test */
     public function it_will_remove_files_when_deleting_a_media_object_with_a_custom_path_generator()
     {
-        config(['medialibrary.path_generator' => TestPathGenerator::class]);
+        config(['media-library.path_generator' => TestPathGenerator::class]);
 
         $pathGenerator = new TestPathGenerator();
 
@@ -58,7 +58,7 @@ class DeleteTest extends TestCase
     public function it_will_not_remove_the_files_when_shouldDeletePreservingMedia_returns_true()
     {
         $testModelClass = new class() extends TestModel {
-            public function shouldDeletePreservingMedia()
+            public function shouldDeletePreservingMedia(): bool
             {
                 return true;
             }
@@ -81,7 +81,7 @@ class DeleteTest extends TestCase
     public function it_will_remove_the_files_when_shouldDeletePreservingMedia_returns_false()
     {
         $testModelClass = new class() extends TestModel {
-            public function shouldDeletePreservingMedia()
+            public function shouldDeletePreservingMedia(): bool
             {
                 return false;
             }
