@@ -1,10 +1,10 @@
 <?php
 
-namespace Spatie\MediaLibrary\Tests\Feature\Models\Media;
+namespace Spatie\MediaLibrary\Tests\Feature\Media;
 
 use Carbon\Carbon;
-use Spatie\MediaLibrary\Exceptions\InvalidConversion;
-use Spatie\MediaLibrary\Exceptions\UrlCannotBeDetermined;
+use RuntimeException;
+use Spatie\MediaLibrary\MediaCollections\Exceptions\InvalidConversion;
 use Spatie\MediaLibrary\Tests\TestCase;
 
 class GetUrlTest extends TestCase
@@ -60,7 +60,7 @@ class GetUrlTest extends TestCase
     {
         $media = $this->testModelWithConversion->addMedia($this->getTestJpg())->toMediaCollection();
 
-        $this->expectException(UrlCannotBeDetermined::class);
+        $this->expectException(RuntimeException::class);
 
         $media->getTemporaryUrl(Carbon::now()->addMinutes(5));
     }
