@@ -1,9 +1,9 @@
 <?php
 
-namespace Spatie\MediaLibrary\Tests\Feature\Models\Media;
+namespace Spatie\MediaLibrary\Tests\Feature\Media;
 
-use Spatie\MediaLibrary\Tests\Support\TestModels\TestModel;
 use Spatie\MediaLibrary\Tests\TestCase;
+use Spatie\MediaLibrary\Tests\TestSupport\TestModels\TestModel;
 
 class MoveTest extends TestCase
 {
@@ -25,7 +25,7 @@ class MoveTest extends TestCase
         $movedMedia = $media->move($anotherModel, 'images');
 
         $this->assertCount(0, $model->getMedia('default'));
-        $this->assertFileNotExists($this->getMediaDirectory($media->id.'/test.jpg'));
+        $this->assertFileDoesNotExist($this->getMediaDirectory($media->id.'/test.jpg'));
 
         $this->assertCount(1, $anotherModel->getMedia('images'));
         $this->assertFileExists($this->getMediaDirectory($movedMedia->id.'/test.jpg'));
@@ -54,7 +54,7 @@ class MoveTest extends TestCase
         $movedMedia = $media->move($anotherModel, 'images', $diskName);
 
         $this->assertCount(0, $model->getMedia('default'));
-        $this->assertFileNotExists($this->getMediaDirectory($media->id.'/test.jpg'));
+        $this->assertFileDoesNotExist($this->getMediaDirectory($media->id.'/test.jpg'));
 
         $this->assertCount(1, $anotherModel->getMedia('images'));
         $this->assertFileExists($this->getTempDirectory('media2').'/'.$movedMedia->id.'/test.jpg');
